+++ src/main/java/com/smplatform/product_service/domain/product/repository/ThumbnailRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ThumbnailRepository extends JpaRepository<Thumbnail, Long> {
-    List<Thumbnail> findAllByProduct_Id(int productId);
+    List<Thumbnail> findAllByProduct_Id(long productId);
 }
